+++ src/routes/(pages)/groups/+page.js
@@ -2,7 +2,6 @@ export function load() {
 	return {
 		title: 'Groups',
 		caption: 'Image by Jule Verne on Pixabay',
-		image: 'groups',
 		static: true,
 	};
 }