+++ src/routes/(pages)/what-we-believe/+page.js
@@ -2,5 +2,6 @@ export function load() {
 	return {
 		title: 'What we believe',
 		caption: 'Image by Gerd Altmann on Pixabay',
+		static: true,
 	};
 }