+++ src/routes/(pages)/where-we-are/+page.js
@@ -2,5 +2,6 @@ export function load() {
 	return {
 		title: 'Where we are',
 		caption: 'Image by Ylanite Koppens on Pixabay',
+		static: true,
 	};
 }