+++ changelog.md
@@ -20,7 +20,7 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Fixed
 
--	added static flag for static pages
+-	fixed two "charity" links not opening in separate windows
 
 ## [2.4.0] - 2025-07-03