+++ changelog.md
@@ -12,8 +12,7 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Added
 
-- add page "Who's who"
-- add additional council members
+-
 
 ### Changed
 
@@ -21,8 +20,14 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Fixed
 
--	fixed two "charity" links not opening in separate windows
+-
 
+## [2.5.0] - 2025-07-06
+
+### Added
+
+- add page "Who's who"
+- add additional council members
 ## [2.4.0] - 2025-07-03
 
 ### Added